worker/requirements.txt
google-cloud-pubsub
tensorflow